+++ src/agent.cpp
@@ -21,9 +21,9 @@ auto agent::ask_phase(uint_t t, ask_fn a, permit_public_status_fn s, int seed) -
   interface_->ask_phase(t, std::move(a), std::move(s), seed);
 }
 
-auto agent::on_bought(const region& s, uint_t t, value_t v) -> void { interface_->on_bought(s, t, v); }
+auto agent::on_bought(region_view s, uint_t t, value_t v) -> void { interface_->on_bought(s, t, v); }
 
-auto agent::on_sold(const region& s, uint_t t, value_t v) -> void { interface_->on_sold(s, t, v); }
+auto agent::on_sold(region_view s, uint_t t, value_t v) -> void { interface_->on_sold(s, t, v); }
 
 auto agent::stop(uint_t t, int seed) -> bool { return interface_->stop(t, seed); }
